+++ lib/src/refs/provider/future_provider.dart
@@ -46,8 +46,8 @@ class FutureProvider<T> extends AsyncRefWidget<T> {
   final Future<T>? future;
 
   @Deprecated('Use `context.read/watch` instead.')
-  static T of<T>(BuildContext context, {bool listen = true, Object? key}) {
-    return listen ? context.watch(key: key) : context.read(key: key);
+  static T of<T>(BuildContext context, {bool listen = true}) {
+    return listen ? context.watch() : context.read();
   }
 
   @override